libmedia: add pre jb-4.2.1 signature for AudioSystem::getDeviceConnectionState
With jb-4.2.1 the type audio_devices_t found in system/core/include/system/audio.h
was redefined from a typedef enum to uint32_t.  This causes the signature
of AudioSystem::getDeviceConnectionState to change in libmedia.so.
Any older than 4.2.1 prebuilt audio.primary.___.so binaries (such as mine from ICS)
may refer to the old signature.  This patch adds back in that reference.

Change-Id: Ie4f92eaec20d581c9bebc805cfd25f8558406e30
Signed-off-by: Hashcode <hashcode0f@gmail.com>